requirements.txt
aiavatar==0.8.13
auraloss>=0.4
chardet
cn2an
einops
eunjeon
fastapi
faster-whisper
ffmpeg-python
funasr
g2p_en
g2pk2
gradio
gruut
huggingface-hub
inflect
jamo
jieba
jieba-fast
ko-pron
librosa>=0.10,<0.11
litellm
matplotlib<3.9
modelscope
nltk
nnAudio

# Torch 系（CUDA に合わせて調整）
torch
torchaudio
torchmetrics
pytorch-lightning

onnxruntime-gpu
openai-whisper
opencc
packaging

# ❗ pandas は 3.x 禁止
pandas>=2.0,<3

peft
pesq
Pillow
pydantic>=2,<3
pypinyin
pyqt5
pytesseract
python-dotenv
regex
requests
rotary-embedding-torch
scipy
sounddevice
soundfile
spaces
split-lang
ToJyutping
tqdm
transformers
typeguard
ultralytics
uvicorn
wordsegment
x-transformers

# Windows 用
pywin32
